event progress + script fixes
implemented the game and event progress using the manager from start of the game to up to the 3rd event. also fixed and edited the PhoneRings, CryingGirl, ClosetOpensSlightly and ShutDoor Script to get rid of exploits and bugs in event progression and added functionality so its not a problem no matter what the player does to the Closet Doors before the event happens. also added creeking door sound for the ClosetOpensSlightly Script.